Major additions and changes - OMGWTF-level
+ vProto.BaseClient and vProto.BaseServer now both inherit from
vProto.Internals._RequestHandler - a class with methods for adding
request handling functions to specific request types;
+ Added vProto.Collections.RequestHandlerCollection - a collection to
store, manage and invoke handlers for specific request types;
+ Added internal requests with dedicated handlers and collections to
vProto.BaseClient;
+ vProto.BaseServer now measures speed as well, from all connected
clients;
+ Added a class library project for common code of the test server and
client;
+ Added synchronous RMI (Remote Method Invocation) through internal
requests, remoting proxies and magic;
> Renamed namespace vProto.Internal_Utilities to vProto.Internals;
> Pending request and response collections are now public with internal
constructors;
> Whether a vProto.BaseClient handles a client for a server is now
determined by the presence of an owning vProto.BaseServer;
> Ping is now divided by 2 to account for both in- and out-going ping -
it approximates the one-way delay this way;
> Updated the description of the vProto.BaseClient.Ping property to
reflect the truth about the property - it is in no way an accurate and
trustable value, it is only an approximation;
> Time taken to fully receive a package is now stored into the package -
perhaps for future use;
> A sending thread is now used per client connection - this is the only
way to avoid the error that was happening;
> Properly and accurately measure outgoing speed by using a buffer to
write to the input stream;
> No longer subtracting ping from the timeout of an incoming request -
this has a huge chance to mess up;
> Sending back-buffer is now cleaned before becoming the front-buffer
for sending - this means the locked section only switches variables,
resulting in significant speed improvement;
> Now the sending uses a back-queue as well, which is cleaned before
becoming the front-queue - swapping also occurs when the buffers are
swapped.

Test program windows now display incoming and outgoing speeds. Client
also displays ping.
Test client sends sends an RMI now when typing in a non-integer string
instead of a plain inline request.

RMI supports returning values and exceptions. Support for out parameters
is not yet ironed out. More experimentation is required.